toio-pet-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toio-pet-agent"
version = "0.1.0"
description = "虚实融合桌宠助手 - Toio + AI Agent + Desktop Pet"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ethen Yang"}
]

dependencies = [
    # BLE 通信
    "bleak>=0.21.0",
    # AI / LLM
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.1.0",
    "ollama>=0.2.0",
    "openai>=1.30.0",
    # 异步 & 通信
    "websockets>=12.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # 图像处理
    "Pillow>=10.3.0",
    # 数据处理
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
    # 工具
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
]

[project.scripts]
toio-pet = "toio_pet_agent.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 100
target-version = ["py310"]
